Added file handling and improved user search

+++ User.h
@@ -10,7 +10,7 @@ class User
 public:
     string get_name() const; //Returning user's name
     long long get_id() const;        //Returning user's ID
-    int get_age() const;     //Returning user's age
+    size_t get_age() const;     //Returning user's age
 
     void display_user() const; //Method that displays User's data name,ID,age
